+++ src/themes/five/src/components/Wallets/index.js
@@ -22,7 +22,7 @@ import {
 import { WalletTransactionItem } from '../WalletTransactionItem'
 import { NotFoundSource } from '../../../../../components/NotFoundSource'
 import { Tabs, Tab } from '../../styles/Tabs'
-import { GiftCard } from '../GiftCard'
+import { GiftCardUI } from '../GiftCard/GiftCardUI'
 
 const WalletsUI = (props) => {
   const {
@@ -103,7 +103,7 @@ const WalletsUI = (props) => {
                   </span>
                 </BalanceElement>
               </SectionWrapper>
-              <GiftCard />
+              <GiftCardUI />
               <div className='transactions_list'>
                 {!transactionsList?.loading && transactionsList.list?.[`wallet:${currentWalletSelected?.id}`]?.length > 0 && (
                   <>
